+++ Backend/src/account.js
@@ -1,4 +1,4 @@
-import { UserData, PostData, CommentData } from "./Database/models/DB_Schemas.js";
+import { UserData, PostData, CommentData } from "../Database/models/DB_Schemas.js";
 import { BackendErrorType } from "./BackendError.js";
 
 // This file contains the functions that handle user account creation and login.
@@ -11,7 +11,7 @@ import { BackendErrorType } from "./BackendError.js";
  * @returns {JSON} - An object containing the result of the login operation and user account information.
  * @throws {BackendErrorType} - Throws an error if the provided query is invalid or if the login credentials are incorrect.
  */
-export async function login(username, password){
+export async function login(username, password) {
     let result = await UserData.findOne({ username: username, password: password });
 
     if (result == null)
@@ -33,25 +33,25 @@ export async function login(username, password){
  *                              or if there's an error during user creation.
 
  */
-export async function signup(firstname, lastname, username, password, email){
-        let result = await UserData.findOne({ username: username })
+export async function signup(firstname, lastname, username, password, email) {
+    let result = await UserData.findOne({ username: username })
 
-        if (result != null)
-            throw BackendErrorType.USERNAME_EXISTS;
+    if (result != null)
+        throw BackendErrorType.USERNAME_EXISTS;
 
-        // create a new user
-        const user = await UserData.create({
-            user_first_name: firstname,
-            user_last_name: lastname,
-            user_email: email,
-            username: username,
-            password: password,
-            user_bio: "",
-            user_following_list: [],
-            user_follower_list: [],
-            user_post_list: [],
-            date_created: new Date(),
-        });
+    // create a new user
+    const user = await UserData.create({
+        user_first_name: firstname,
+        user_last_name: lastname,
+        user_email: email,
+        username: username,
+        password: password,
+        user_bio: "",
+        user_following_list: [],
+        user_follower_list: [],
+        user_post_list: [],
+        date_created: new Date(),
+    });
 
-        return { result: 'SUCCESS', message: 'New Account Created' };
+    return { result: 'SUCCESS', message: 'New Account Created' };
 }
